requirements.txt
python-binance==1.0.10
discord.py==1.7.2
